refactor:Refactored Update page
